+++ tensorcircuit/interfaces/tensortrans.py
@@ -46,6 +46,10 @@ def tensor_to_dtype(t: Tensor) -> str:
     return which_backend(t).dtype(t)  # type: ignore
 
 
+def tensor_to_dlpack(t: Tensor) -> Any:
+    return which_backend(t).to_dlpack(t)
+
+
 def general_args_to_numpy(args: Any) -> Any:
     """
     Given a pytree, get the corresponding numpy array pytree
@@ -90,3 +94,28 @@ def numpy_args_to_backend(
         t = backend.tree_map(target_backend.convert_to_tensor, args)
         t = backend.tree_map(target_backend.cast, t, dtype)
         return t
+
+
+def general_args_to_backend(
+    args: Any, dtype: Any = None, target_backend: Any = None, enable_dlpack: bool = True
+) -> Any:
+    if not enable_dlpack:
+        # TODO(@refraction-ray): add device shift for numpy mediate transformation
+        args = general_args_to_numpy(args)
+        args = numpy_args_to_backend(args, dtype, target_backend)
+        return args
+
+    caps = backend.tree_map(tensor_to_dlpack, args)
+    if target_backend is None:
+        target_backend = backend
+    elif isinstance(target_backend, str):
+        target_backend = get_backend(target_backend)
+    if dtype is None:
+        return backend.tree_map(target_backend.from_dlpack, caps)
+    if isinstance(dtype, str):
+        leaves, treedef = backend.tree_flatten(args)
+        dtype = [dtype for _ in range(len(leaves))]
+        dtype = backend.tree_unflatten(treedef, dtype)
+    t = backend.tree_map(target_backend.from_dlpack, caps)
+    t = backend.tree_map(target_backend.cast, t, dtype)
+    return t